feat: major layout refactors and fixes
