+++ src/core/utils/mod.rs
@@ -4,6 +4,7 @@ pub mod dialogs;
 pub mod files;
 pub mod fonts;
 pub mod sorting;
+pub mod tabs;
 pub mod text;
 pub mod widgets;
 
@@ -12,5 +13,6 @@ pub use colors::*;
 pub use dialogs::*;
 pub use files::*;
 pub use sorting::*;
+pub use tabs::*;
 pub use text::*;
 pub use widgets::*;
