+++ src/gui/windows/containers/tabs.rs
@@ -1,62 +1,13 @@
-use crate::core::fs::MY_PC_PATH;
-use crate::core::portable;
 use crate::gui::i18n::I18n;
-use crate::gui::icons::IconCache;
 use crate::gui::theme::ThemePalette;
-use crate::gui::utils::{
-    clear_clipboard_files, clickable_icon, expand_environment_variables, truncate_item_text,
-};
-use crate::gui::windows::containers::enums::TabbarNavAction;
-use crate::gui::windows::containers::structs::{TabInfo, TabView, TabbarAction, TabsAction};
+use crate::gui::utils::{clickable_icon, truncate_item_text};
+use crate::gui::windows::containers::structs::{TabInfo, TabsAction};
 use crate::gui::windows::windowsoverrides::handle_draw_windows_buttons;
 use eframe::egui;
 use egui::{FontFamily, FontId};
 use egui_phosphor::{fill, regular};
 use std::path::{Path, PathBuf};
 use windows::Win32::Foundation::HWND;
-
-fn nav_icon_button(
-    ui: &mut egui::Ui,
-
-    icon: &str,
-    palette: &ThemePalette,
-    enabled: bool,
-    hover_text: &str,
-) -> egui::Response {
-    let font_id = egui::FontId::default();
-    let resp = ui.add_enabled(
-        enabled,
-        egui::Label::new(
-            egui::RichText::new(icon)
-                .font(font_id.clone())
-                .color(ui.visuals().text_color()),
-        )
-        .selectable(false)
-        .sense(egui::Sense::click()),
-    );
-
-    if enabled && resp.hovered() {
-        ui.painter().text(
-            resp.rect.center(),
-            egui::Align2::CENTER_CENTER,
-            icon,
-            font_id.clone(),
-            palette.primary,
-        );
-    }
-
-    let resp = resp.on_hover_text(
-        egui::RichText::new(hover_text)
-            .size(palette.tooltip_text_size)
-            .color(palette.tooltip_text_color),
-    );
-
-    if enabled {
-        resp.on_hover_cursor(egui::CursorIcon::PointingHand)
-    } else {
-        resp
-    }
-}
 
 pub fn draw_tabs(
     ui: &mut egui::Ui,
@@ -84,13 +35,11 @@ pub fn draw_tabs(
         egui::vec2(ui.available_width(), 32.0),
         egui::Layout::left_to_right(egui::Align::Center),
         |ui| {
-            // --- TABS AREA (CLIPPED) ---
             ui.allocate_ui_with_layout(
                 egui::vec2(tabs_width, 32.0),
                 egui::Layout::left_to_right(egui::Align::Min),
                 |ui| {
                     let tabs_rect = ui.available_rect_before_wrap();
-                    // 🔥 THIS is where overflow must be handled
                     egui::ScrollArea::horizontal()
                         .id_salt("tabs_scroll")
                         .auto_shrink([false, true])
@@ -98,7 +47,6 @@ pub fn draw_tabs(
                         .show(ui, |ui| {
                             ui.horizontal(|ui| {
                                 for tab in tabs {
-                                    // Compute tab width dynamically
                                     let tab_width = 140.0;
                                     let (rect, resp) = ui.allocate_exact_size(
                                         egui::vec2(tab_width, 28.0),
@@ -494,569 +442,4 @@ fn tab_add_button(
     }
 
     resp
-}
-
-fn toolbar_buttons(
-    ui: &mut egui::Ui,
-    i18n: &I18n,
-    palette: &ThemePalette,
-    is_favorited: bool,
-    is_root: bool,
-    can_go_back: bool,
-    can_go_forward: bool,
-) -> TabbarAction {
-    let mut action = TabbarAction::default();
-
-    // Navigation buttons
-    if nav_icon_button(
-        ui,
-        regular::ARROW_LEFT,
-        palette,
-        can_go_back,
-        &i18n.tr("tooltip_nav_back"),
-    )
-    .clicked()
-        && can_go_back
-    {
-        action.nav = Some(TabbarNavAction::Back);
-    }
-
-    if nav_icon_button(
-        ui,
-        regular::ARROW_RIGHT,
-        palette,
-        can_go_forward,
-        &i18n.tr("tooltip_nav_forward"),
-    )
-    .clicked()
-        && can_go_forward
-    {
-        action.nav = Some(TabbarNavAction::Forward);
-    }
-
-    let can_go_up = !is_root;
-    if nav_icon_button(
-        ui,
-        regular::ARROW_UP,
-        palette,
-        can_go_up,
-        &i18n.tr("tooltip_nav_up"),
-    )
-    .clicked()
-        && can_go_up
-    {
-        action.nav = Some(TabbarNavAction::Up);
-    }
-
-    if clickable_icon(ui, regular::ARROWS_CLOCKWISE, palette.primary)
-        .on_hover_text(
-            egui::RichText::new(i18n.tr("tooltip_refresh"))
-                .size(palette.tooltip_text_size)
-                .color(palette.tooltip_text_color),
-        )
-        .on_hover_cursor(egui::CursorIcon::PointingHand)
-        .clicked()
-    {
-        action.refresh_current_directory = true;
-        clear_clipboard_files();
-    }
-
-    // Action buttons
-    if clickable_icon(ui, regular::FOLDER_PLUS, palette.primary)
-        .on_hover_text(
-            egui::RichText::new(i18n.tr("tooltip_newfolder"))
-                .size(palette.tooltip_text_size)
-                .color(palette.tooltip_text_color),
-        )
-        .on_hover_cursor(egui::CursorIcon::PointingHand)
-        .clicked()
-    {
-        action.create_folder = true;
-    }
-
-    if clickable_icon(ui, regular::FILE_PLUS, palette.primary)
-        .on_hover_text(
-            egui::RichText::new(i18n.tr("tooltip_newfile"))
-                .size(palette.tooltip_text_size)
-                .color(palette.tooltip_text_color),
-        )
-        .on_hover_cursor(egui::CursorIcon::PointingHand)
-        .clicked()
-    {
-        action.create_file = true;
-    }
-
-    let star_icon = if is_favorited {
-        fill::STAR
-    } else {
-        regular::STAR
-    };
-    let star_color = if is_favorited {
-        palette.button_favorite_fill
-    } else {
-        palette.tooltip_text_color
-    };
-
-    let star_font = if is_favorited {
-        FontId::new(palette.text_size, FontFamily::Name("phosphor_fill".into()))
-    } else {
-        FontId::new(palette.text_size, FontFamily::Proportional)
-    };
-
-    let star_resp = ui.add_enabled(
-        !is_root,
-        egui::Label::new(
-            egui::RichText::new(star_icon)
-                .font(star_font)
-                .color(star_color),
-        )
-        .selectable(false)
-        .sense(egui::Sense::click()),
-    );
-
-    let star_resp = if is_root {
-        star_resp.on_hover_text(
-            egui::RichText::new(i18n.tr("tooltip_favorites_disabled"))
-                .size(palette.tooltip_text_size)
-                .color(palette.tooltip_text_color),
-        )
-    } else {
-        star_resp
-            .on_hover_text(
-                egui::RichText::new(if is_favorited {
-                    i18n.tr("tooltip_favorites_remove")
-                } else {
-                    i18n.tr("tooltip_favorites_add")
-                })
-                .size(palette.tooltip_text_size)
-                .color(palette.tooltip_text_color),
-            )
-            .on_hover_cursor(egui::CursorIcon::PointingHand)
-    };
-
-    if star_resp.clicked() {
-        if is_favorited {
-            action.remove_favorite = true;
-        } else {
-            action.add_favorite = true;
-        }
-    }
-
-    action
-}
-
-pub fn draw_tabbar(
-    ui: &mut egui::Ui,
-    i18n: &I18n,
-    icon_cache: &IconCache,
-    tab: &mut TabView,
-    tab_id: u64,
-    palette: &ThemePalette,
-    is_favorited: bool,
-    drag_active: bool,
-    drag_hover_target: Option<PathBuf>,
-) -> TabbarAction {
-    let mut action = TabbarAction::default();
-    let tabbar_rect = ui.available_rect_before_wrap();
-    let pointer_pos = ui.input(|i| i.pointer.interact_pos().or_else(|| i.pointer.hover_pos()));
-    let pointer_released =
-        ui.input(|i| i.pointer.any_released() && i.pointer.interact_pos().is_some());
-    let hovered_target_ref = drag_hover_target.as_ref();
-    let mut breadcrumb_drop_target: Option<PathBuf> = None;
-    let pointer_in_tabbar = pointer_pos
-        .map(|pos| tabbar_rect.contains(pos))
-        .unwrap_or(false);
-    let can_go_back = tab.nav.can_go_back();
-    let can_go_forward = tab.nav.can_go_forward();
-
-    ui.horizontal(|ui| {
-        let toolbar_action = toolbar_buttons(
-            ui,
-            i18n,
-            palette,
-            is_favorited,
-            tab.nav.is_root(),
-            can_go_back,
-            can_go_forward,
-        );
-
-        // Merge toolbar actions
-        if toolbar_action.nav.is_some() {
-            action.nav = toolbar_action.nav;
-        }
-        if toolbar_action.refresh_current_directory {
-            action.refresh_current_directory = true;
-        }
-        if toolbar_action.create_folder {
-            action.create_folder = true;
-        }
-        if toolbar_action.create_file {
-            action.create_file = true;
-        }
-        if toolbar_action.add_favorite {
-            action.add_favorite = true;
-        }
-        if toolbar_action.remove_favorite {
-            action.remove_favorite = true;
-        }
-
-        ui.separator();
-
-        // 👇 SWITCH BETWEEN MODES
-        if tab.breadcrumb_path_editing {
-            let text_edit_id = ui.id().with(("breadcrumbs_path_edit", tab_id));
-
-            // --- 🔥 Shake animation ---
-            let mut offset_x = 0.0;
-            if tab.breadcrumb_path_error {
-                let t = (ui.input(|i| i.time) - tab.breadcrumb_path_error_animation_time) as f32;
-
-                if t < 0.4 {
-                    let frequency = 30.0_f32;
-                    let amplitude = 4.0 * (1.0 - t / 0.4); // decay
-                    offset_x = (t * frequency).sin() * amplitude;
-                }
-            }
-
-            ui.add_space(offset_x);
-
-            let time_since_error = ui.input(|i| i.time) - tab.breadcrumb_path_error_animation_time;
-            let error_strength = (1.0 - time_since_error * 2.0).clamp(0.0, 1.0);
-
-            let stroke_color = if tab.breadcrumb_path_error {
-                egui::Color32::from_rgba_premultiplied(255, 80, 80, (255.0 * error_strength) as u8)
-            } else {
-                ui.visuals().widgets.inactive.bg_stroke.color
-            };
-
-            let frame = egui::Frame::NONE
-                .stroke(egui::Stroke::new(1.5, stroke_color))
-                .corner_radius(egui::CornerRadius::same(4));
-
-            let mut resp = frame
-                .show(ui, |ui| {
-                    ui.add(
-                        egui::TextEdit::singleline(&mut tab.breadcrumb_path_buffer)
-                            .id(text_edit_id)
-                            .frame(!tab.breadcrumb_path_error)
-                            .desired_width(ui.available_width() - 40.0)
-                            .font(FontId::new(
-                                palette.text_size,
-                                egui::FontFamily::Proportional,
-                            )),
-                    )
-                })
-                .inner;
-
-            if resp.changed() {
-                tab.breadcrumb_path_error = false;
-            }
-
-            if tab.breadcrumb_path_error && resp.hovered() {
-                resp = resp.on_hover_text(
-                    egui::RichText::new(i18n.tr("tooltip_path_does_not_exist"))
-                        .size(palette.tooltip_text_size)
-                        .color(palette.tooltip_text_color),
-                );
-            }
-
-            if !tab.breadcrumb_just_started_editing || tab.breadcrumb_path_error {
-                resp.request_focus();
-                tab.breadcrumb_just_started_editing = true;
-            }
-
-            // ✅ Track focus for global shortcut blocking
-            action.is_breadcrumb_path_edit_active = resp.has_focus();
-
-            // ✅ Handle input behavior (DON’T REMOVE THIS)
-            let enter = ui.input(|i| i.key_pressed(egui::Key::Enter));
-            let escape = ui.input(|i| i.key_pressed(egui::Key::Escape));
-
-            let mut exit_edit_mode = false;
-
-            if enter {
-                let input = tab.breadcrumb_path_buffer.trim().trim_matches('"');
-
-                // Expand environment variables before checking path existence
-                let expanded_input = expand_environment_variables(input);
-                let new_path = PathBuf::from(&expanded_input);
-
-                if new_path.exists() {
-                    action.nav_to = Some(new_path);
-                    exit_edit_mode = true;
-                } else {
-                    println!(
-                        "{}: {} ({}: {})",
-                        i18n.tr("tooltip_invalid_path"),
-                        tab.breadcrumb_path_buffer,
-                        i18n.tr("tooltip_invalid_path_expanded"),
-                        expanded_input
-                    );
-                    tab.breadcrumb_path_error = true;
-                    tab.breadcrumb_path_error_animation_time = ui.input(|i| i.time);
-                }
-            } else if escape || resp.lost_focus() {
-                tab.breadcrumb_path_buffer = tab.nav.current.to_string_lossy().to_string();
-                exit_edit_mode = true;
-            }
-
-            if exit_edit_mode {
-                tab.breadcrumb_path_editing = false;
-                tab.breadcrumb_just_started_editing = false;
-
-                // 🔥 Reset error state
-                tab.breadcrumb_path_error = false;
-                tab.breadcrumb_path_error_animation_time = 0.0;
-
-                ui.memory_mut(|mem| mem.surrender_focus(text_edit_id));
-            }
-        } else if tab.nav.is_root() {
-            let pc_icon_path = PathBuf::from("C:\\");
-            if let Some(icon) = icon_cache.get(&pc_icon_path, true) {
-                ui.add(egui::Image::new(&icon).fit_to_exact_size(egui::vec2(14.0, 14.0)));
-            }
-
-            if ui
-                .add(
-                    egui::Label::new(
-                        egui::RichText::new(i18n.tr("thispc"))
-                            .size(palette.text_size)
-                            .color(palette.text_header_section),
-                    )
-                    .selectable(false)
-                    .sense(egui::Sense::click()),
-                )
-                .clicked()
-            {
-                action.nav_to = Some(PathBuf::from(MY_PC_PATH));
-            }
-        } else {
-            let segments =
-                build_breadcrumbs(&tab.nav.current, ui.available_width(), palette.text_size);
-            let mut first = true;
-
-            let mut breadcrumbs_right = 0.0;
-            let font_id = egui::FontId::new(palette.text_size, egui::FontFamily::Proportional);
-            let total_width = ui.available_width();
-
-            for (label, path) in segments.iter() {
-                if !first {
-                    let old_spacing = ui.spacing().item_spacing;
-                    ui.spacing_mut().item_spacing.x = 10.0;
-
-                    ui.label(
-                        egui::RichText::new(">")
-                            .size(palette.text_size)
-                            .color(palette.text_header_section),
-                    );
-
-                    ui.spacing_mut().item_spacing = old_spacing;
-                }
-                first = false;
-
-                let inner = egui::Frame::NONE
-                    .fill(egui::Color32::TRANSPARENT)
-                    .inner_margin(egui::Margin::symmetric(6, 2))
-                    .corner_radius(egui::CornerRadius::same(palette.medium_radius))
-                    .show(ui, |ui| {
-                        let color = ui.visuals().text_color();
-
-                        let max_width = (total_width / segments.len() as f32).clamp(60.0, 180.0);
-
-                        let (label_text, truncated) =
-                            truncate_item_text(ui, label, max_width, &font_id, color);
-
-                        let resp = ui.add(
-                            egui::Label::new(
-                                egui::RichText::new(label_text)
-                                    .size(palette.text_size)
-                                    .color(palette.text_header_section),
-                            )
-                            .selectable(false)
-                            .sense(egui::Sense::click()),
-                        );
-
-                        // ✅ RETURN THIS
-                        if truncated {
-                            resp.on_hover_text(label)
-                        } else {
-                            resp
-                        }
-                    });
-
-                let resp = inner.response.union(inner.inner);
-
-                if drag_active {
-                    let breadcrumb_target_path = if label == "..." {
-                        segments.first().map(|(_, root_path)| root_path)
-                    } else {
-                        Some(path)
-                    };
-                    let hovered = breadcrumb_target_path
-                        .and_then(|target_path| {
-                            hovered_target_ref.map(|target| target == target_path)
-                        })
-                        .unwrap_or_else(|| {
-                            pointer_pos
-                                .map(|pointer| resp.rect.contains(pointer))
-                                .unwrap_or(false)
-                        });
-                    if hovered {
-                        let painter = ui
-                            .ctx()
-                            .layer_painter(egui::LayerId::new(
-                                egui::Order::Background,
-                                ui.id().with("breadcrumb_drop_bg").with(path),
-                            ))
-                            .with_clip_rect(ui.clip_rect());
-                        painter.rect_filled(
-                            resp.rect,
-                            egui::CornerRadius::same(palette.medium_radius),
-                            palette.primary_hover,
-                        );
-
-                        if pointer_released {
-                            if label == "..." {
-                                if let Some((_, root_path)) = segments.first() {
-                                    breadcrumb_drop_target = Some(root_path.clone());
-                                    action.move_files_to_breadcrumb_dir_rect = Some(resp.rect);
-                                }
-                            } else {
-                                breadcrumb_drop_target = Some(path.clone());
-                                action.move_files_to_breadcrumb_dir_rect = Some(resp.rect);
-                            }
-                        }
-                    }
-                }
-
-                if resp.hovered() {
-                    ui.ctx().set_cursor_icon(egui::CursorIcon::PointingHand);
-                }
-
-                if !drag_active && resp.clicked() {
-                    // If "..." clicked, navigate to root
-                    if label == "..." {
-                        if let Some((_, root_path)) = segments.first() {
-                            action.nav_to = Some(root_path.clone());
-                        }
-                    } else {
-                        action.nav_to = Some(path.clone());
-                    }
-                }
-
-                breadcrumbs_right = resp.rect.right();
-            }
-
-            // ---- Detect click in empty area to enter path editing ----
-            let available_rect = ui.available_rect_before_wrap();
-            let empty_area = egui::Rect::from_min_max(
-                egui::pos2(breadcrumbs_right, available_rect.top()),
-                available_rect.right_bottom(),
-            );
-
-            if ui
-                .interact(
-                    empty_area,
-                    ui.id().with("breadcrumb_empty"),
-                    egui::Sense::click(),
-                )
-                .clicked()
-            {
-                tab.breadcrumb_path_editing = true;
-                tab.breadcrumb_path_buffer = tab.nav.current.to_string_lossy().to_string();
-            }
-        }
-    });
-
-    // Mouse side buttons for history navigation.
-    if action.nav.is_none() && pointer_in_tabbar && !tab.breadcrumb_path_editing {
-        if can_go_back && ui.input(|i| i.pointer.button_pressed(egui::PointerButton::Extra1)) {
-            action.nav = Some(TabbarNavAction::Back);
-        } else if can_go_forward
-            && ui.input(|i| i.pointer.button_pressed(egui::PointerButton::Extra2))
-        {
-            action.nav = Some(TabbarNavAction::Forward);
-        }
-    }
-
-    action.move_files_to_breadcrumb_dir = breadcrumb_drop_target;
-    action
-}
-
-fn build_breadcrumbs(path: &Path, available_width: f32, font_size: f32) -> Vec<(String, PathBuf)> {
-    use std::path::Component;
-
-    // Approximate width per character
-    let char_width = font_size * 0.55;
-    let separator_width = char_width * 4.0; // width for '>' separator
-
-    // Collect all segments
-    let all_segments = if portable::is_portable_path(&path.to_path_buf()) {
-        portable::build_breadcrumb_segments(&path.to_path_buf()).unwrap_or_default()
-    } else {
-        let mut segments = Vec::new();
-        let mut current = PathBuf::new();
-        for comp in path.components() {
-            match comp {
-                Component::Prefix(prefix) => {
-                    current.push(prefix.as_os_str());
-                    segments.push((
-                        prefix.as_os_str().to_string_lossy().to_string(),
-                        current.clone(),
-                    ));
-                }
-                Component::RootDir => {
-                    current.push(Path::new("\\"));
-                }
-                Component::Normal(name) => {
-                    current.push(name);
-                    segments.push((name.to_string_lossy().to_string(), current.clone()));
-                }
-                _ => {}
-            }
-        }
-        segments
-    };
-
-    // Compute total width
-    let total_width: f32 = all_segments
-        .iter()
-        .map(|(label, _)| label.len() as f32 * char_width + separator_width)
-        .sum();
-
-    // If everything fits, show all segments
-    if total_width <= available_width {
-        return all_segments;
-    }
-
-    // Otherwise, truncate dynamically
-    let mut segments = Vec::new();
-
-    // Always show "..." pointing to root
-    if let Some((_, root_path)) = all_segments.first() {
-        segments.push(("...".to_string(), root_path.clone()));
-    }
-
-    // Start from the end and add segments until we fill the available space
-    let mut used_width = 0.0;
-    let mut included_segments = Vec::new();
-
-    // Reverse iterator for trailing segments
-    for (label, path) in all_segments.iter().rev() {
-        let label_width = label.len() as f32 * char_width + separator_width;
-        if used_width + label_width > available_width - char_width * 3.0 {
-            // Leave room for "..." and a little buffer
-            break;
-        }
-        used_width += label_width;
-        included_segments.push((label.clone(), path.clone()));
-    }
-
-    // Reverse to keep the correct order
-    included_segments.reverse();
-
-    // Append trailing segments after "..."
-    segments.extend(included_segments);
-
-    segments
 }
